+++ packages/eyes-api/src/input/Configuration.ts
@@ -104,7 +104,8 @@ export type Configuration<TElement = unknown, TSelector = unknown> = GeneralConf
   VGConfiguration
 
 export class ConfigurationData<TElement = unknown, TSelector = unknown>
-  implements Required<Configuration<TElement, TSelector>> {
+  implements Required<Configuration<TElement, TSelector>>
+{
   protected readonly _spec: ConfigurationSpec<TElement, TSelector>
 
   private _config: Configuration<TElement, TSelector> = {}
@@ -512,7 +513,7 @@ export class ConfigurationData<TElement = unknown, TSelector = unknown>
     this._config.properties = properties
   }
   getProperties(): PropertyDataData[] {
-    return this.properties.map(property => new PropertyDataData(property))
+    return this.properties?.map(property => new PropertyDataData(property)) ?? []
   }
   setProperties(properties: PropertyData[]): this {
     this.properties = properties
@@ -522,6 +523,7 @@ export class ConfigurationData<TElement = unknown, TSelector = unknown>
   addProperty(prop: PropertyData): this
   addProperty(propOrName: PropertyData | string, value?: string): this {
     const property = utils.types.isString(propOrName) ? {name: propOrName, value} : propOrName
+    if (!this.properties) this.properties = []
     this.properties.push(property)
     return this
   }