+++ packages/eyes-api/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- fix missing `properties` array in `Configuration`
 
 ## 1.0.1 - 2021/5/12